Modify cargo dev command of dioxus project

+++ dioxus/src/bin/dev.rs
@@ -8,7 +8,9 @@ use std::path::Path;
 
 fn main() {
     let target_dx_folder_path = Path::new(&current_dir().unwrap()).join("target").join("dx");
-    fs::remove_dir_all(target_dx_folder_path).unwrap();
+    if target_dx_folder_path.exists() {
+        fs::remove_dir_all(target_dx_folder_path).unwrap();
+    }
     let port = 3000;
     if is_port_in_use(port) {
         eprintln!("Port {} is already in use.", port);
