+++ dioxus/src/bin/make.rs
@@ -5,7 +5,9 @@ use std::path::Path;
 
 fn main() {
     let target_dx_folder_path = Path::new(&current_dir().unwrap()).join("target").join("dx");
-    fs::remove_dir_all(target_dx_folder_path).unwrap();
+    if target_dx_folder_path.exists() {
+        fs::remove_dir_all(target_dx_folder_path).unwrap();
+    }
     let _ = Command::new("dx")
         .args(["bundle", "--release", "--web"])
         .current_dir(current_dir().unwrap())
